requirements.txt
narwhals==1.26.0
numpy==2.2.2
packaging==24.2
pandas==2.2.3
pillow==11.1.0
plotly==6.0.0
python-dateutil==2.9.0.post0
pytz==2025.1
scipy==1.15.1
six==1.17.0
tzdata==2025.1
